rewrite tests for the adapter classes
this will allow the tests to have access to private variables
